+++ db/migrate/20260323160918_add_metadata_to_test_runs.rb
@@ -1,0 +1,5 @@
+class AddMetadataToTestRuns < ActiveRecord::Migration[8.1]
+  def change
+    add_column :test_runs, :metadata, :jsonb, default: {}, null: false
+  end
+end